[backend] Split playbook components into separated files (#14283)
