+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-components.ts
@@ -13,101 +13,54 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 */
 import * as R from 'ramda';
-import { v4 as uuidv4 } from 'uuid';
 import type { JSONSchemaType } from 'ajv';
-import * as jsonpatch from 'fast-json-patch';
 import { type BasicStoreEntityPlaybook, ENTITY_TYPE_PLAYBOOK, type PlaybookComponent } from './playbook-types';
-import {
-  type AuthorizedMember,
-  AUTOMATION_MANAGER_USER,
-  AUTOMATION_MANAGER_USER_UUID,
-  executionContext,
-  isUserCanAccessStixElement,
-  isUserInPlatformOrganization,
-  SYSTEM_USER,
-} from '../../utils/access';
+import { AUTOMATION_MANAGER_USER, AUTOMATION_MANAGER_USER_UUID, executionContext, isUserCanAccessStixElement, isUserInPlatformOrganization, SYSTEM_USER } from '../../utils/access';
 import { pushToConnector, pushToWorkerForConnector } from '../../database/rabbitmq';
-import {
-  ABSTRACT_STIX_CORE_OBJECT,
-  ABSTRACT_STIX_CORE_RELATIONSHIP,
-  ABSTRACT_STIX_CYBER_OBSERVABLE,
-  ABSTRACT_STIX_DOMAIN_OBJECT,
-  ABSTRACT_STIX_RELATIONSHIP,
-  ENTITY_TYPE_CONTAINER,
-  ENTITY_TYPE_THREAT_ACTOR,
-  INPUT_ASSIGNEE,
-  INPUT_AUTHORIZED_MEMBERS,
-  INPUT_CREATED_BY,
-  INPUT_GRANTED_REFS,
-  INPUT_KILLCHAIN,
-  INPUT_LABELS,
-  INPUT_MARKINGS,
-  INPUT_PARTICIPANT,
-  OPENCTI_ADMIN_UUID,
-} from '../../schema/general';
-import type { BasicStoreCommon, BasicStoreRelation, StoreCommon, StoreRelation } from '../../types/store';
-import { generateInternalId, generateStandardId, idGenFromData } from '../../schema/identifier';
-import { now, observableValue, utcDate } from '../../utils/format';
+import { ABSTRACT_STIX_CORE_RELATIONSHIP, ABSTRACT_STIX_CYBER_OBSERVABLE, ENTITY_TYPE_CONTAINER } from '../../schema/general';
+import type { BasicStoreRelation, StoreRelation } from '../../types/store';
+import { generateInternalId, generateStandardId } from '../../schema/identifier';
+import { utcDate } from '../../utils/format';
 import type { StixCampaign, StixContainer, StixIncident, StixInfrastructure, StixMalware, StixReport, StixThreatActor } from '../../types/stix-2-1-sdo';
 import { convertStixToInternalTypes, generateInternalType, getParentTypes } from '../../schema/schemaUtils';
-import {
-  ENTITY_TYPE_ATTACK_PATTERN,
-  ENTITY_TYPE_CAMPAIGN,
-  ENTITY_TYPE_CONTAINER_REPORT,
-  ENTITY_TYPE_INCIDENT,
-  ENTITY_TYPE_INTRUSION_SET,
-  ENTITY_TYPE_MALWARE,
-  ENTITY_TYPE_TOOL,
-  isStixDomainObjectContainer,
-  STIX_DOMAIN_OBJECT_CONTAINER_CASES,
-} from '../../schema/stixDomainObject';
-import type { CyberObjectExtension, StixBundle, StixCoreObject, StixCyberObject, StixDomainObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
-import { STIX_EXT_MITRE, STIX_EXT_OCTI, STIX_EXT_OCTI_SCO } from '../../types/stix-2-1-extensions';
+import { ENTITY_TYPE_CONTAINER_REPORT, isStixDomainObjectContainer } from '../../schema/stixDomainObject';
+import type { CyberObjectExtension, StixBundle, StixCyberObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
+import { STIX_EXT_OCTI, STIX_EXT_OCTI_SCO } from '../../types/stix-2-1-extensions';
 import { connectorsForPlaybook } from '../../database/repository';
-import { internalFindByIds, fullEntitiesList, fullRelationsList, storeLoadById } from '../../database/middleware-loader';
-import { type BasicStoreEntityOrganization, ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../organization/organization-types';
-import { getEntitiesMapFromCache, getEntityFromCache } from '../../database/cache';
-import { createdBy, objectLabel, objectMarking } from '../../schema/stixRefRelationship';
+import { fullEntitiesList, fullRelationsList, storeLoadById } from '../../database/middleware-loader';
+import { getEntityFromCache } from '../../database/cache';
 import { logApp } from '../../config/conf';
-import { FunctionalError } from '../../config/errors';
-import { extractStixRepresentative } from '../../database/stix-representative';
 import { isEmptyField, isNotEmptyField, READ_RELATIONSHIPS_INDICES, READ_RELATIONSHIPS_INDICES_WITHOUT_INFERRED } from '../../database/utils';
-import { schemaAttributesDefinition } from '../../schema/schema-attributes';
-import { schemaRelationsRefDefinition } from '../../schema/schema-relationsRef';
 import { stixLoadByIds } from '../../database/middleware';
 import { usableNotifiers } from '../notifier/notifier-domain';
 import { convertToNotificationUser, type DigestEvent, EVENT_NOTIFICATION_VERSION } from '../../manager/notificationManager';
 import { storeNotificationEvent } from '../../database/stream/stream-handler';
 import { ENTITY_TYPE_SETTINGS } from '../../schema/internalObject';
 import { isStixCyberObservable } from '../../schema/stixCyberObservable';
-import { createStixPattern } from '../../python/pythonBridge';
-import { generateKeyValueForIndicator } from '../../domain/stixCyberObservable';
-import { RELATION_BASED_ON, RELATION_INDICATES } from '../../schema/stixCoreRelationship';
+import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
 import type { StixRelation } from '../../types/stix-2-1-sro';
-import { extractValidObservablesFromIndicatorPattern, STIX_PATTERN_TYPE } from '../../utils/syntax';
-import { ENTITY_TYPE_CONTAINER_CASE_INCIDENT, type StixCaseIncident } from '../case/case-incident/case-incident-types';
 import { isStixMatchFilterGroup } from '../../utils/filtering/filtering-stix/stix-filtering';
-import { ENTITY_TYPE_INDICATOR, type StixIndicator } from '../indicator/indicator-types';
-import { ENTITY_TYPE_CONTAINER_CASE_RFI } from '../case/case-rfi/case-rfi-types';
-import { ENTITY_TYPE_CONTAINER_CASE_RFT } from '../case/case-rft/case-rft-types';
+import { ENTITY_TYPE_INDICATOR } from '../indicator/indicator-types';
 import { ENTITY_TYPE_CONTAINER_TASK, type StixTask, type StoreEntityTask } from '../task/task-types';
-import { EditOperation, FilterMode } from '../../generated/graphql';
-import { ENTITY_TYPE_MARKING_DEFINITION } from '../../schema/stixMetaObject';
-import { schemaTypesDefinition } from '../../schema/schema-types';
+import { FilterMode } from '../../generated/graphql';
 import { generateCreateMessage } from '../../database/data-changes';
-import { ENTITY_TYPE_CONTAINER_CASE } from '../case/case-types';
 import { findAllByCaseTemplateId } from '../task/task-domain';
 import type { BasicStoreEntityTaskTemplate } from '../task/task-template/task-template-types';
 import type { BasicStoreSettings } from '../../types/settings';
-import { AUTHORIZED_MEMBERS_SUPPORTED_ENTITY_TYPES, buildRestrictedMembers } from '../../utils/authorizedMembers';
-import { ENTITY_TYPE_CONTAINER_GROUPING } from '../grouping/grouping-types';
-import { ENTITY_TYPE_CONTAINER_FEEDBACK } from '../case/feedback/feedback-types';
 import { PLAYBOOK_SEND_EMAIL_TEMPLATE_COMPONENT } from './components/send-email-template-component';
-import { applyOperationFieldPatch, convertMembersToUsers, extractBundleBaseElement } from './playbook-utils';
+import { convertMembersToUsers, extractBundleBaseElement } from './playbook-utils';
 import { PLAYBOOK_DATA_STREAM_PIR } from './components/data-stream-pir-component';
 import { convertStoreToStix_2_1 } from '../../database/stix-2-1-converter';
-import { ENTITY_TYPE_SECURITY_COVERAGE, INPUT_COVERED, type StixSecurityCoverage, type StoreEntitySecurityCoverage } from '../securityCoverage/securityCoverage-types';
 import { pushAll } from '../../utils/arrayUtil';
+import { PLAYBOOK_CONTAINER_WRAPPER_COMPONENT } from './components/container-wrapper-component';
+import { PLAYBOOK_MANIPULATE_KNOWLEDGE_COMPONENT } from './components/manipulate-knowledge-component';
+import { PLAYBOOK_SECURITY_COVERAGE_COMPONENT } from './components/security-coverage-component';
+import { PLAYBOOK_SHARING_COMPONENT } from './components/sharing-component';
+import { PLAYBOOK_UNSHARING_COMPONENT } from './components/unsharing-component';
+import { PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT } from './components/access-restrictions-component';
+import { PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT } from './components/remove-access-restrictions-component';
+import { PLAYBOOK_CREATE_INDICATOR_COMPONENT } from './components/create-indicator-component';
+import { PLAYBOOK_CREATE_OBSERVABLE_COMPONENT } from './components/create-observable-component';
 
 // region built in playbook components
 interface LoggerConfiguration {
@@ -460,45 +413,6 @@ export const PLAYBOOK_CONNECTOR_COMPONENT: PlaybookComponent<ConnectorConfigurat
   },
 };
 
-interface ContainerWrapperConfiguration {
-  container_type: string;
-  caseTemplates: { label: string; value: string }[];
-  all: boolean;
-  excludeMainElement: boolean;
-  copyFiles: boolean;
-  newContainer: boolean;
-}
-const PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_SCHEMA: JSONSchemaType<ContainerWrapperConfiguration> = {
-  type: 'object',
-  properties: {
-    container_type: { type: 'string', $ref: 'Container type', default: '', oneOf: [] },
-    caseTemplates: {
-      type: 'array',
-      uniqueItems: true,
-      default: [],
-      $ref: 'Case templates',
-      items: { type: 'string', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Wrap all elements included in the bundle', default: false },
-    excludeMainElement: { type: 'boolean', $ref: 'Exclude main element from container', default: false },
-    copyFiles: { type: 'boolean', $ref: 'Copy files from main element to the container', default: false },
-    newContainer: { type: 'boolean', $ref: 'Create a new container at each run', default: false },
-  },
-  required: ['container_type'],
-};
-
-// For now, only a fixed list of containers are compatible
-// these are the containers that can be created with a name and no specific mandatory fields
-const PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_AVAILABLE_CONTAINERS = [
-  ENTITY_TYPE_CONTAINER_REPORT,
-  ENTITY_TYPE_CONTAINER_GROUPING,
-  ENTITY_TYPE_CONTAINER_CASE_INCIDENT,
-  ENTITY_TYPE_CONTAINER_CASE_RFI,
-  ENTITY_TYPE_CONTAINER_CASE_RFT,
-  ENTITY_TYPE_CONTAINER_FEEDBACK,
-  ENTITY_TYPE_CONTAINER_TASK,
-];
-
 export const buildStixTaskFromTaskTemplate = (taskTemplate: BasicStoreEntityTaskTemplate, container: StixContainer) => {
   const taskData = {
     name: taskTemplate.name,
@@ -532,739 +446,6 @@ export const createTaskFromCaseTemplates = async (
     }
   }
   return tasks;
-};
-
-export const PLAYBOOK_CONTAINER_WRAPPER_COMPONENT: PlaybookComponent<ContainerWrapperConfiguration> = {
-  id: 'PLAYBOOK_CONTAINER_WRAPPER_COMPONENT',
-  name: 'Container wrapper',
-  description: 'Create a container and wrap the element inside it',
-  icon: 'container',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_SCHEMA,
-  schema: async () => {
-    const elements = PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_AVAILABLE_CONTAINERS.map((t) => ({ const: t, title: t }));
-    const schemaElement = { properties: { container_type: { oneOf: elements } } };
-    return R.mergeDeepRight<JSONSchemaType<ContainerWrapperConfiguration>, any>(PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_SCHEMA, schemaElement);
-  },
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const { container_type, all, excludeMainElement, copyFiles, newContainer, caseTemplates } = playbookNode.configuration;
-    if (!PLAYBOOK_CONTAINER_WRAPPER_COMPONENT_AVAILABLE_CONTAINERS.includes(container_type)) {
-      throw FunctionalError('this container type is incompatible with the Container Wrapper playbook component', { container_type });
-    }
-    if (container_type) {
-      const baseData = extractBundleBaseElement(dataInstanceId, bundle);
-      const created = newContainer ? now() : baseData.extensions[STIX_EXT_OCTI].created_at;
-      const representative = extractStixRepresentative(baseData);
-      let name = `Generated container wrapper from playbook at ${created}`;
-      if (representative && newContainer) {
-        name = `${representative} - ${created}`;
-      } else if (representative) {
-        name = representative;
-      }
-      const containerData: Record<string, unknown> = {
-        name,
-        created,
-      };
-      if (container_type === ENTITY_TYPE_CONTAINER_REPORT) {
-        containerData.published = created;
-      }
-      if (container_type === ENTITY_TYPE_CONTAINER_GROUPING) {
-        containerData.context = 'playbook';
-      }
-      const standardId = generateStandardId(container_type, containerData);
-      const storeContainer = {
-        standard_id: standardId,
-        entity_type: container_type,
-        parent_types: getParentTypes(container_type),
-        ...containerData,
-      } as StoreCommon;
-      const container = convertStoreToStix_2_1(storeContainer) as StixReport | StixCaseIncident;
-      // add all objects in the container if requested in the playbook config
-      if (all) {
-        // If excludeMainElement is true and all is true, exclude the main element from the container
-        if (excludeMainElement) {
-          container.object_refs = bundle.objects.filter((o: StixObject) => o.id !== baseData.id).map((o: StixObject) => o.id);
-        } else {
-          container.object_refs = bundle.objects.map((o: StixObject) => o.id);
-        }
-      } else {
-        container.object_refs = [baseData.id];
-      }
-      // Specific remapping of some attributes, waiting for a complete binding solution in the UI
-      // Following attributes are the same as the base instance: description, content, markings, labels, created_by, assignees, participants
-      if ((baseData as StixReport).description) {
-        container.description = (baseData as StixReport).description;
-      }
-      if ((baseData as StixReport).extensions[STIX_EXT_OCTI].content) {
-        (container as StixReport).extensions[STIX_EXT_OCTI].content = (baseData as StixReport).extensions[STIX_EXT_OCTI].content;
-      }
-      if ((baseData as StixCaseIncident).content) {
-        (container as StixCaseIncident).content = (baseData as StixCaseIncident).content;
-      }
-      if (baseData.object_marking_refs) {
-        container.object_marking_refs = baseData.object_marking_refs;
-      }
-      if ((<StixDomainObject>baseData).labels) {
-        container.labels = (<StixDomainObject>baseData).labels;
-      }
-      if ((<StixDomainObject>baseData).created_by_ref) {
-        container.created_by_ref = (<StixDomainObject>baseData).created_by_ref;
-      }
-      if (baseData.extensions[STIX_EXT_OCTI].participant_ids) {
-        container.extensions[STIX_EXT_OCTI].participant_ids = baseData.extensions[STIX_EXT_OCTI].participant_ids;
-      }
-      if (baseData.extensions[STIX_EXT_OCTI].assignee_ids) {
-        container.extensions[STIX_EXT_OCTI].assignee_ids = baseData.extensions[STIX_EXT_OCTI].assignee_ids;
-      }
-      // if the base instance is an incident and we wrap into an Incident Case, we set the same severity
-      if ((<StixIncident>baseData) && container_type === ENTITY_TYPE_CONTAINER_CASE_INCIDENT) {
-        (<StixCaseIncident>container).severity = (<StixIncident>baseData).severity;
-        (<StixCaseIncident>container).external_references = (<StixIncident>baseData).external_references;
-        (<StixCaseIncident>container).extensions[STIX_EXT_OCTI].granted_refs = (<StixIncident>baseData).extensions[STIX_EXT_OCTI].granted_refs;
-      }
-      // Copy files from the main element to the container if requested
-      if (copyFiles && baseData.extensions[STIX_EXT_OCTI].files && baseData.extensions[STIX_EXT_OCTI].files.length > 0) {
-        container.extensions[STIX_EXT_OCTI].files = baseData.extensions[STIX_EXT_OCTI].files;
-      }
-      if (STIX_DOMAIN_OBJECT_CONTAINER_CASES.includes(container_type) && caseTemplates.length > 0) {
-        const tasks = await createTaskFromCaseTemplates(caseTemplates, (container as StixContainer));
-        pushAll(bundle.objects, tasks);
-      }
-      bundle.objects.push(container);
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-
-interface SecurityCoverageConfiguration {
-  all: boolean;
-  auto_enrichment_disable: boolean;
-  periodicity: string;
-  duration: string;
-  type_affinity: string;
-  platforms_affinity: string[];
-}
-const PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA: JSONSchemaType<SecurityCoverageConfiguration> = {
-  type: 'object',
-  properties: {
-    all: { type: 'boolean', $ref: 'Create a security coverage for each element of the bundle (on compatible types)', default: false },
-    auto_enrichment_disable: { type: 'boolean', $ref: 'Force manual coverage (prevent enrichment connectors from running)', default: false },
-    periodicity: { type: 'string', $ref: 'Coverage recurrence (every x)', default: 'P1D' },
-    duration: { type: 'string', $ref: 'Duration', default: 'P30D' },
-    type_affinity: {
-      type: 'string',
-      $ref: 'Type affinity',
-      default: 'ENDPOINT',
-    },
-    platforms_affinity: {
-      type: 'array',
-      uniqueItems: true,
-      default: ['windows', 'linux', 'macos'],
-      $ref: 'Platform(s) affinity',
-      items: { type: 'string', oneOf: [] },
-    },
-  },
-  required: ['periodicity', 'duration', 'type_affinity', 'platforms_affinity'],
-};
-
-const SECURITY_COVERAGE_COMPATIBLE_TYPES = [
-  'report',
-  'grouping',
-  'case-incident',
-  'x-opencti-case-incident',
-  'intrusion-set',
-  'campaign',
-  'incident',
-];
-
-export const PLAYBOOK_SECURITY_COVERAGE_COMPONENT: PlaybookComponent<SecurityCoverageConfiguration> = {
-  id: 'PLAYBOOK_SECURITY_COVERAGE_COMPONENT',
-  name: 'Security coverage',
-  description: 'Create a security coverage for the given entity(ies) (when type is compatible)',
-  icon: 'security-coverage',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_SECURITY_COVERAGE_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const { all, auto_enrichment_disable, periodicity, duration, type_affinity, platforms_affinity } = playbookNode.configuration;
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle) as StixDomainObject;
-    if (SECURITY_COVERAGE_COMPATIBLE_TYPES.includes(baseData.type)) {
-      const name = extractStixRepresentative(baseData);
-      const securityCoverageData: Record<string, unknown> = {
-        name,
-        created: now(),
-        auto_enrichment_disable: auto_enrichment_disable,
-        periodicity: periodicity,
-        duration: duration,
-        type_affinity: type_affinity,
-        platforms_affinity: platforms_affinity,
-        [INPUT_COVERED]: { standard_id: baseData.id },
-        [INPUT_LABELS]: (baseData.labels ?? []).map((l) => ({ value: l })),
-      };
-      const standardId = generateStandardId(ENTITY_TYPE_SECURITY_COVERAGE, securityCoverageData);
-      const storeSecurityCoverage = {
-        internal_id: uuidv4(),
-        standard_id: standardId,
-        entity_type: ENTITY_TYPE_SECURITY_COVERAGE,
-        parent_types: getParentTypes(ENTITY_TYPE_SECURITY_COVERAGE),
-        ...securityCoverageData,
-      } as StoreEntitySecurityCoverage;
-      const securityCoverage = convertStoreToStix_2_1(storeSecurityCoverage) as StixSecurityCoverage;
-      bundle.objects.push(securityCoverage);
-    }
-    if (all) {
-      for (let index = 0; index < bundle.objects.length; index += 1) {
-        const element = bundle.objects[index] as StixDomainObject;
-        if (SECURITY_COVERAGE_COMPATIBLE_TYPES.includes(element.type)) {
-          const name = extractStixRepresentative(element);
-          const securityCoverageData: Record<string, unknown> = {
-            name,
-            created: now(),
-            auto_enrichment_disable: auto_enrichment_disable,
-            periodicity: periodicity,
-            duration: duration,
-            type_affinity: type_affinity,
-            [INPUT_COVERED]: { standard_id: element.id },
-            [INPUT_LABELS]: (element.labels ?? []).map((l) => ({ value: l })),
-          };
-          const standardId = generateStandardId(ENTITY_TYPE_SECURITY_COVERAGE, securityCoverageData);
-          const storeContainer = {
-            internal_id: uuidv4(),
-            standard_id: standardId,
-            entity_type: ENTITY_TYPE_SECURITY_COVERAGE,
-            parent_types: getParentTypes(ENTITY_TYPE_SECURITY_COVERAGE),
-            ...securityCoverageData,
-          } as StoreCommon;
-          const securityCoverage = convertStoreToStix_2_1(storeContainer) as StixSecurityCoverage;
-          bundle.objects.push(securityCoverage);
-        }
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-
-export interface SharingConfiguration {
-  organizations: string[] | { label: string; value: string }[];
-  all: boolean;
-}
-const PLAYBOOK_SHARING_COMPONENT_SCHEMA: JSONSchemaType<SharingConfiguration> = {
-  type: 'object',
-  properties: {
-    organizations: {
-      type: 'array',
-      uniqueItems: true,
-      default: [],
-      $ref: 'Target organizations',
-      items: { type: 'string', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Share all elements included in the bundle', default: false },
-  },
-  required: ['organizations'],
-};
-export const PLAYBOOK_SHARING_COMPONENT: PlaybookComponent<SharingConfiguration> = {
-  id: 'PLAYBOOK_SHARING_COMPONENT',
-  name: 'Share with organizations',
-  description: 'Share with organizations within the platform',
-  icon: 'organization-add',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_SHARING_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_SHARING_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const context = executionContext('playbook_components');
-    const { organizations, all } = playbookNode.configuration;
-    const organizationsValues = organizations.map((o) => (typeof o !== 'string' ? o.value : o));
-    const organizationsByIds = await internalFindByIds<BasicStoreEntityOrganization>(context, SYSTEM_USER, organizationsValues, {
-      type: ENTITY_TYPE_IDENTITY_ORGANIZATION,
-      baseData: true,
-      baseFields: ['standard_id'],
-    }) as BasicStoreEntityOrganization[];
-    if (organizationsByIds.length === 0) {
-      return { output_port: 'out', bundle }; // nothing to do since organizations are empty
-    }
-    const organizationIds = organizationsByIds.map((o) => o.standard_id);
-    for (let index = 0; index < bundle.objects.length; index += 1) {
-      const element = bundle.objects[index];
-      if (all || element.id === dataInstanceId) {
-        element.extensions[STIX_EXT_OCTI].granted_refs = [...(element.extensions[STIX_EXT_OCTI].granted_refs ?? []), ...organizationIds];
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-
-export interface UnsharingConfiguration {
-  organizations: string[] | { label: string; value: string }[];
-  all: boolean;
-}
-const PLAYBOOK_UNSHARING_COMPONENT_SCHEMA: JSONSchemaType<UnsharingConfiguration> = {
-  type: 'object',
-  properties: {
-    organizations: {
-      type: 'array',
-      uniqueItems: true,
-      default: [],
-      $ref: 'Target organizations',
-      items: { type: 'string', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Unshare all elements included in the bundle', default: false },
-  },
-  required: ['organizations'],
-};
-export const PLAYBOOK_UNSHARING_COMPONENT: PlaybookComponent<UnsharingConfiguration> = {
-  id: 'PLAYBOOK_UNSHARING_COMPONENT',
-  name: 'Unshare with organizations',
-  description: 'Unshare with organizations within the platform',
-  icon: 'organization-remove',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_UNSHARING_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_UNSHARING_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const context = executionContext('playbook_components', AUTOMATION_MANAGER_USER);
-    const { organizations, all } = playbookNode.configuration;
-    const organizationsValues = organizations.map((o) => (typeof o !== 'string' ? o.value : o));
-    const organizationsByIds = await internalFindByIds<BasicStoreEntityOrganization>(context, SYSTEM_USER, organizationsValues, {
-      type: ENTITY_TYPE_IDENTITY_ORGANIZATION,
-      baseData: true,
-      baseFields: ['standard_id'],
-    }) as BasicStoreEntityOrganization[];
-    if (organizationsByIds.length === 0) {
-      return { output_port: 'out', bundle }; // nothing to do since organizations are empty
-    }
-    const organizationIds = organizationsByIds.map((o) => o.standard_id);
-    const patchOperations = [];
-    for (let index = 0; index < bundle.objects.length; index += 1) {
-      const element = bundle.objects[index];
-      if (all || element.id === dataInstanceId) {
-        const patchValue = {
-          op: EditOperation.Remove,
-          path: `/objects/${index}/extensions/${STIX_EXT_OCTI}/granted_refs`,
-          value: organizationIds,
-        };
-        const patchOperation = {
-          operation: patchValue.op,
-          key: INPUT_GRANTED_REFS,
-          value: patchValue.value,
-        };
-        applyOperationFieldPatch(element, [patchOperation]);
-        patchOperations.push(patchValue);
-      }
-    }
-    if (patchOperations.length > 0) {
-      const patchedBundle = jsonpatch.applyPatch(structuredClone(bundle), patchOperations).newDocument;
-      const diff = jsonpatch.compare(bundle, patchedBundle);
-      if (isNotEmptyField(diff)) {
-        return { output_port: 'out', bundle: patchedBundle };
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-
-export interface AccessRestrictionsConfiguration {
-  access_restrictions: { groupsRestriction: { label: string; value: string; type: string }[]; accessRight: string; label: string; type: string; value: string }[];
-  all: boolean;
-}
-const PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA: JSONSchemaType<AccessRestrictionsConfiguration> = {
-  type: 'object',
-  properties: {
-    access_restrictions: {
-      type: 'array',
-      uniqueItems: true,
-      default: [{
-        label: 'Administrator',
-        type: 'User',
-        value: OPENCTI_ADMIN_UUID,
-        accessRight: 'admin',
-        groupsRestriction: [],
-      }],
-      $ref: 'Access restrictions',
-      items: { type: 'object', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Apply access restrictions on all elements included in the bundle', default: false },
-  },
-  required: ['access_restrictions'],
-};
-export const PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT: PlaybookComponent<AccessRestrictionsConfiguration> = {
-  id: 'PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT',
-  name: 'Manage access restrictions',
-  description: 'Manage advanced access restrictions on entities',
-  icon: 'lock',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const context = executionContext('playbook_components');
-    const { access_restrictions: accessRestrictions, all } = playbookNode.configuration;
-    // Resolve potential dynamic access rights
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle) as StixObject;
-    const finalAccessRestrictions = [];
-    for (let index = 0; index < accessRestrictions.length; index += 1) {
-      const accessRestriction = accessRestrictions[index];
-      if (accessRestriction.value === 'AUTHOR') {
-        // If dynamic binding of author and an author is really defined in the data
-        const createdById = baseData.extensions[STIX_EXT_OCTI].created_by_ref_id;
-        const createdByType = baseData.extensions[STIX_EXT_OCTI].created_by_ref_type;
-        if (isNotEmptyField(createdById) && createdByType === ENTITY_TYPE_IDENTITY_ORGANIZATION) {
-          finalAccessRestrictions.push({ ...accessRestriction, value: createdById });
-        }
-      } else if (accessRestriction.value === 'CREATORS') {
-        const creators = (baseData.extensions[STIX_EXT_OCTI].creator_ids ?? []).filter((id) => isNotEmptyField(id));
-        for (let index2 = 0; index2 < creators.length; index2 += 1) {
-          finalAccessRestrictions.push({ ...accessRestriction, value: creators[index2] });
-        }
-      } else if (accessRestriction.value === 'ASSIGNEES') {
-        const assignees = (baseData.extensions[STIX_EXT_OCTI].assignee_ids ?? []).filter((id) => isNotEmptyField(id));
-        for (let index2 = 0; index2 < assignees.length; index2 += 1) {
-          finalAccessRestrictions.push({ ...accessRestriction, value: assignees[index2] });
-        }
-      } else if (accessRestriction.value === 'PARTICIPANTS') {
-        const participants = (baseData.extensions[STIX_EXT_OCTI].participant_ids ?? []).filter((id) => isNotEmptyField(id));
-        for (let index2 = 0; index2 < participants.length; index2 += 1) {
-          finalAccessRestrictions.push({ ...accessRestriction, value: participants[index2] });
-        }
-      } else if (accessRestriction.value === 'BUNDLE_ORGANIZATIONS') {
-        const bundleOrganizations = bundle.objects.filter((o) => o.extensions[STIX_EXT_OCTI].type === ENTITY_TYPE_IDENTITY_ORGANIZATION);
-        const bundleOrganizationsIds = bundleOrganizations.map((o) => o.extensions[STIX_EXT_OCTI].id).filter((id) => isNotEmptyField(id));
-        for (let index2 = 0; index2 < bundleOrganizationsIds.length; index2 += 1) {
-          finalAccessRestrictions.push({ ...accessRestriction, value: bundleOrganizationsIds[index2] });
-        }
-      } else {
-        finalAccessRestrictions.push(accessRestriction);
-      }
-    }
-    const patchOperations = [];
-    const input = finalAccessRestrictions.map((n) => ({
-      id: n.value,
-      access_right: n.accessRight,
-      groups_restriction_ids: n.groupsRestriction.map((o) => o.value),
-    }));
-    if (input.length === 0) {
-      return { output_port: 'out', bundle };
-    }
-    for (let index = 0; index < bundle.objects.length; index += 1) {
-      const element = bundle.objects[index];
-      const internalType = generateInternalType(element);
-      if (AUTHORIZED_MEMBERS_SUPPORTED_ENTITY_TYPES.includes(internalType) && (all || element.id === dataInstanceId)) {
-        const args = {
-          entityId: element.id,
-          input,
-          requiredCapabilities: ['KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS'],
-          entityType: internalType,
-          busTopicKey: ABSTRACT_STIX_DOMAIN_OBJECT,
-        };
-        const restrictedMembers = await buildRestrictedMembers(context, AUTOMATION_MANAGER_USER, args) as AuthorizedMember[];
-        const patchValue = {
-          op: EditOperation.Replace,
-          path: `/objects/${index}/extensions/${STIX_EXT_OCTI}/authorized_members`,
-          value: restrictedMembers,
-        };
-        const patchOperation = {
-          operation: EditOperation.Replace,
-          key: INPUT_AUTHORIZED_MEMBERS,
-          value: restrictedMembers,
-        };
-        applyOperationFieldPatch(element, [patchOperation]);
-        patchOperations.push(patchValue);
-      }
-    }
-    if (patchOperations.length > 0) {
-      const patchedBundle = jsonpatch.applyPatch(structuredClone(bundle), patchOperations).newDocument;
-      const diff = jsonpatch.compare(bundle, patchedBundle);
-      if (isNotEmptyField(diff)) {
-        return { output_port: 'out', bundle: patchedBundle };
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-export interface RemoveAccessRestrictionsConfiguration {
-  all: boolean;
-}
-const PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA: JSONSchemaType<RemoveAccessRestrictionsConfiguration> = {
-  type: 'object',
-  properties: {
-    all: { type: 'boolean', $ref: 'Remove access restrictions on all elements included in the bundle', default: false },
-  },
-  required: [],
-};
-export const PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT: PlaybookComponent<RemoveAccessRestrictionsConfiguration> = {
-  id: 'PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT',
-  name: 'Remove access restrictions',
-  description: 'Remove advanced access restrictions on entities',
-  icon: 'lock-remove',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }],
-  configuration_schema: PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const { all } = playbookNode.configuration;
-    const patchOperations = [];
-    for (let index = 0; index < bundle.objects.length; index += 1) {
-      const element = bundle.objects[index];
-      const internalType = generateInternalType(element);
-      if (AUTHORIZED_MEMBERS_SUPPORTED_ENTITY_TYPES.includes(internalType) && (all || element.id === dataInstanceId)) {
-        const patchValue = {
-          op: EditOperation.Replace,
-          path: `/objects/${index}/extensions/${STIX_EXT_OCTI}/restricted_members`,
-          value: [],
-        };
-        const patchOperation = {
-          operation: EditOperation.Replace,
-          key: INPUT_AUTHORIZED_MEMBERS,
-          value: [],
-        };
-        applyOperationFieldPatch(element, [patchOperation]);
-        patchOperations.push(patchValue);
-      }
-    }
-    if (patchOperations.length > 0) {
-      const patchedBundle = jsonpatch.applyPatch(structuredClone(bundle), patchOperations).newDocument;
-      const diff = jsonpatch.compare(bundle, patchedBundle);
-      if (isNotEmptyField(diff)) {
-        return { output_port: 'out', bundle: patchedBundle };
-      }
-    }
-    return { output_port: 'out', bundle };
-  },
-};
-
-const attributePathMapping: any = {
-  [INPUT_MARKINGS]: {
-    [ABSTRACT_STIX_CORE_OBJECT]: `/${objectMarking.stixName}`,
-    [ABSTRACT_STIX_RELATIONSHIP]: `/${objectMarking.stixName}`,
-  },
-  [INPUT_LABELS]: {
-    [ABSTRACT_STIX_CORE_OBJECT]: `/${objectLabel.stixName}`,
-    [ABSTRACT_STIX_RELATIONSHIP]: `/${objectLabel.stixName}`,
-  },
-  [INPUT_CREATED_BY]: {
-    [ABSTRACT_STIX_CORE_OBJECT]: `/${createdBy.stixName}`,
-    [ABSTRACT_STIX_RELATIONSHIP]: `/${createdBy.stixName}`,
-  },
-  [INPUT_ASSIGNEE]: {
-    [ABSTRACT_STIX_DOMAIN_OBJECT]: `/extensions/${STIX_EXT_OCTI}/assignee_ids`,
-  },
-  [INPUT_PARTICIPANT]: {
-    [ABSTRACT_STIX_DOMAIN_OBJECT]: `/extensions/${STIX_EXT_OCTI}/participant_ids`,
-  },
-  confidence: {
-    [ABSTRACT_STIX_DOMAIN_OBJECT]: '/confidence',
-    [ABSTRACT_STIX_RELATIONSHIP]: '/confidence',
-  },
-  x_opencti_score: {
-    [ENTITY_TYPE_INDICATOR]: `/extensions/${STIX_EXT_OCTI}/score`,
-    [ABSTRACT_STIX_CYBER_OBSERVABLE]: `/extensions/${STIX_EXT_OCTI_SCO}/score`,
-    [ENTITY_TYPE_IDENTITY_ORGANIZATION]: `/extensions/${STIX_EXT_OCTI}/score`,
-  },
-  x_opencti_detection: {
-    [ENTITY_TYPE_INDICATOR]: `/extensions/${STIX_EXT_OCTI}/detection`,
-  },
-  x_opencti_workflow_id: {
-    [ABSTRACT_STIX_DOMAIN_OBJECT]: `/extensions/${STIX_EXT_OCTI}/workflow_id`,
-    [ABSTRACT_STIX_CYBER_OBSERVABLE]: `/extensions/${STIX_EXT_OCTI}/workflow_id`,
-    [ABSTRACT_STIX_RELATIONSHIP]: `/extensions/${STIX_EXT_OCTI}/workflow_id`,
-  },
-  severity: {
-    [ENTITY_TYPE_CONTAINER_CASE]: '/severity',
-    [ENTITY_TYPE_INCIDENT]: '/severity',
-  },
-  priority: {
-    [ENTITY_TYPE_CONTAINER_CASE]: '/priority',
-  },
-  indicator_types: {
-    [ENTITY_TYPE_INDICATOR]: '/indicator_types',
-  },
-  [INPUT_KILLCHAIN]: {
-    [ENTITY_TYPE_INDICATOR]: '/kill_chain_phases',
-  },
-  x_mitre_platforms: {
-    [ENTITY_TYPE_INDICATOR]: `/extensions/${STIX_EXT_MITRE}/platforms`,
-  },
-};
-interface UpdateValueConfiguration {
-  label: string;
-  value: string;
-  patch_value: string;
-}
-interface UpdateConfiguration {
-  actions: { op: 'add' | 'replace' | 'remove'; attribute: string; value: UpdateValueConfiguration[] }[];
-  all: boolean;
-}
-const PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT_SCHEMA: JSONSchemaType<UpdateConfiguration> = {
-  type: 'object',
-  properties: {
-    actions: {
-      type: 'array',
-      default: [],
-      items: {
-        type: 'object',
-        properties: {
-          op: { type: 'string', enum: ['add', 'replace', 'remove'] },
-          attribute: { type: 'string' },
-          value: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                label: { type: 'string' },
-                value: { type: 'string' },
-                patch_value: { type: 'string' },
-              },
-              required: ['label', 'value', 'patch_value'],
-            },
-          },
-        },
-        required: ['op', 'attribute', 'value'],
-      },
-    },
-    all: { type: 'boolean', $ref: 'Manipulate all elements included in the bundle', default: false },
-  },
-  required: ['actions'],
-};
-export const PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT: PlaybookComponent<UpdateConfiguration> = {
-  id: 'PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT',
-  name: 'Manipulate knowledge',
-  description: 'Manipulate STIX data',
-  icon: 'edit',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }, { id: 'unmodified', type: 'out' }],
-  configuration_schema: PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT_SCHEMA,
-  executor: async ({ dataInstanceId, playbookNode, bundle }) => {
-    const context = executionContext('playbook_components');
-    const cacheIds = await getEntitiesMapFromCache(context, AUTOMATION_MANAGER_USER, ENTITY_TYPE_MARKING_DEFINITION);
-    const { actions, all } = playbookNode.configuration;
-    // Compute if the attribute is defined as multiple in schema definition
-    const isAttributeMultiple = (entityType: string, attribute: string) => {
-      const baseAttribute = schemaAttributesDefinition.getAttribute(entityType, attribute);
-      if (baseAttribute) return baseAttribute.multiple;
-      const relationRef = schemaRelationsRefDefinition.getRelationRef(entityType, attribute);
-      if (relationRef) return relationRef.multiple;
-      return undefined;
-    };
-    const getAttributeType = (entityType: string, attribute: string) => {
-      const baseAttribute = schemaAttributesDefinition.getAttribute(entityType, attribute);
-      return baseAttribute?.type ?? 'string';
-    };
-    // Compute the access path for the attribute in the static matrix
-    const computeAttributePath = (entityType: string, attribute: string) => {
-      if (attributePathMapping[attribute]) {
-        if (attributePathMapping[attribute][entityType]) {
-          return attributePathMapping[attribute][entityType];
-        }
-        const key = Object.keys(attributePathMapping[attribute]).filter((o) => getParentTypes(entityType).includes(o)).at(0);
-        if (key) {
-          return attributePathMapping[attribute][key];
-        }
-      }
-      return undefined;
-    };
-    const convertValue = (attributeType: string, value: any) => {
-      if (attributeType === 'numeric') return Number(value);
-      if (attributeType === 'boolean') return value.toLowerCase() === 'true';
-      return value;
-    };
-    const patchOperations: jsonpatch.Operation[] = [];
-    for (let index = 0; index < bundle.objects.length; index += 1) {
-      const element = bundle.objects[index];
-      if (all || element.id === dataInstanceId) {
-        const { type, id } = element.extensions[STIX_EXT_OCTI];
-        const elementOperations = actions
-          .map((action) => {
-            const attrPath = computeAttributePath(type, action.attribute);
-            const multiple = isAttributeMultiple(type, action.attribute);
-            const attributeType = getAttributeType(type, action.attribute);
-            return ({ action, multiple, attributeType, attrPath, path: `/objects/${index}${attrPath}` });
-          })
-          // Unrecognized attributes must be filtered
-          .filter(({ attrPath, multiple }) => isNotEmptyField(multiple) && isNotEmptyField(attrPath))
-          // Map actions to data patches
-          .map(({ action, path, multiple, attributeType }) => {
-            if (multiple) {
-              const currentValues = jsonpatch.getValueByPointer(bundle, path) ?? [];
-              // the patch value can be the "label" instead of id (for ex: markings ids / labels ids)
-              const actionPatchValues = action.value.map((o) => {
-                // If value is an id, must be converted to standard_id has we work on stix bundle
-                if (cacheIds.has(o.patch_value)) return (cacheIds.get(o.patch_value) as BasicStoreCommon).standard_id;
-                // Else, just return the value
-                return convertValue(attributeType, o.patch_value);
-              });
-              // the value is always the id
-              const actionValues = action.value.map((o) => {
-                // If value is an id, must be converted to standard_id has we work on stix bundle
-                if (cacheIds.has(o.value)) return (cacheIds.get(o.value) as BasicStoreCommon).standard_id;
-                // Else, just return the value
-                return convertValue(attributeType, o.value);
-              });
-              if (action.op === EditOperation.Add) {
-                return {
-                  op: action.op,
-                  attribute: action.attribute,
-                  value: actionValues,
-                  patchOperation: { op: EditOperation.Replace, path, value: R.uniq([...currentValues, ...actionPatchValues]) },
-                };
-              }
-              if (action.op === EditOperation.Replace) {
-                return {
-                  op: action.op,
-                  attribute: action.attribute,
-                  value: actionValues,
-                  patchOperation: { op: EditOperation.Replace, path, value: actionPatchValues },
-                };
-              }
-              if (action.op === EditOperation.Remove) {
-                return {
-                  op: action.op,
-                  attribute: action.attribute,
-                  value: actionValues,
-                  patchOperation: { op: EditOperation.Replace, path, value: currentValues.filter((c: any) => !actionPatchValues.includes(c)) },
-                };
-              }
-            }
-            const currentPatchValue = R.head(action.value)?.patch_value;
-            const currentValue = R.head(action.value)?.value;
-            return {
-              op: action.op,
-              attribute: action.attribute,
-              value: currentValue,
-              patchOperation: { op: action.op, path, value: convertValue(attributeType, currentPatchValue) },
-            };
-          });
-        // Enlist operations to execute
-        if (elementOperations.length > 0) {
-          const operationObject = elementOperations.map((op) => {
-            return { key: op.attribute, value: Array.isArray(op.value) ? op.value : [op.value], operation: op.op };
-          });
-          if (id) {
-            applyOperationFieldPatch(element, operationObject);
-          }
-          pushAll(patchOperations, elementOperations.map((e) => e.patchOperation));
-        }
-      }
-    }
-    // Apply operations if needed
-    if (patchOperations.length > 0) {
-      const patchedBundle = jsonpatch.applyPatch(structuredClone(bundle), patchOperations).newDocument;
-      const diff = jsonpatch.compare(bundle, patchedBundle);
-      if (isNotEmptyField(diff)) {
-        return { output_port: 'out', bundle: patchedBundle };
-      }
-    }
-    return { output_port: 'unmodified', bundle };
-  },
 };
 
 const DATE_SEEN_RULE = 'seen_dates';
@@ -1506,359 +687,6 @@ const PLAYBOOK_NOTIFIER_COMPONENT: PlaybookComponent<NotifierConfiguration> = {
     return { output_port: undefined, bundle };
   },
 };
-interface CreateIndicatorConfiguration {
-  all: boolean;
-  wrap_in_container: boolean;
-  types: string[];
-}
-const PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA: JSONSchemaType<CreateIndicatorConfiguration> = {
-  type: 'object',
-  properties: {
-    types: {
-      type: 'array',
-      default: [],
-      $ref: 'Types',
-      items: { type: 'string', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Create indicators from all observables in the bundle', default: false },
-    wrap_in_container: { type: 'boolean', $ref: 'If main entity is a container, wrap indicators in container', default: false },
-  },
-  required: [],
-};
-const PLAYBOOK_CREATE_INDICATOR_COMPONENT: PlaybookComponent<CreateIndicatorConfiguration> = {
-  id: 'PLAYBOOK_CREATE_INDICATOR_COMPONENT',
-  name: 'Promote observable to indicator',
-  description: 'Create an indicator based on an observable',
-  icon: 'indicator',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }, { id: 'unmodified', type: 'out' }],
-  configuration_schema: PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA,
-  schema: async () => {
-    const types = schemaTypesDefinition.get(ABSTRACT_STIX_CYBER_OBSERVABLE);
-    const elements = types.map((t) => ({ const: t, title: t }))
-      .sort((a, b) => (a.title.toLowerCase() > b.title.toLowerCase() ? 1 : -1));
-    const schemaElement = { properties: { types: { items: { oneOf: elements } } } };
-    return R.mergeDeepRight<JSONSchemaType<CreateIndicatorConfiguration>, any>(PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA, schemaElement);
-  },
-  executor: async ({ playbookNode, dataInstanceId, bundle }) => {
-    const { all, wrap_in_container, types } = playbookNode.configuration;
-    const context = executionContext('playbook_components');
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle);
-    const observables = [baseData];
-    if (all) {
-      pushAll(observables, bundle.objects);
-    }
-    const { type: baseDataType, id } = baseData.extensions[STIX_EXT_OCTI];
-    const isBaseDataAContainer = isStixDomainObjectContainer(baseDataType);
-    const objectsToPush: StixObject[] = [];
-    for (let index = 0; index < observables.length; index += 1) {
-      const observable = observables[index] as StixCyberObject;
-      let { type } = observable.extensions[STIX_EXT_OCTI];
-      if (isStixCyberObservable(type) && (isEmptyField(types) || types.includes(type))) {
-        const indicatorName = observableValue({ ...observable, entity_type: type });
-        const { key, value } = generateKeyValueForIndicator(type, indicatorName, observable);
-        if (key.includes('Artifact')) {
-          type = 'StixFile';
-        }
-        const pattern = await createStixPattern(context, AUTOMATION_MANAGER_USER, key, value);
-        const score = observable.x_opencti_score ?? observable.extensions[STIX_EXT_OCTI_SCO]?.score;
-        const { granted_refs } = observable.extensions[STIX_EXT_OCTI];
-        if (pattern) {
-          const indicatorData = {
-            name: indicatorName,
-            x_opencti_main_observable_type: type,
-            x_opencti_score: score,
-            pattern,
-            pattern_type: STIX_PATTERN_TYPE,
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: ENTITY_TYPE_INDICATOR,
-                main_observable_type: type,
-                score,
-              },
-            },
-          };
-          const indicatorStandardId = generateStandardId(ENTITY_TYPE_INDICATOR, indicatorData);
-          const storeIndicator = {
-            internal_id: generateInternalId(),
-            standard_id: indicatorStandardId,
-            entity_type: ENTITY_TYPE_INDICATOR,
-            parent_types: getParentTypes(ENTITY_TYPE_INDICATOR),
-            ...indicatorData,
-          } as StoreCommon;
-          const indicator = convertStoreToStix_2_1(storeIndicator) as StixIndicator;
-          if (observable.object_marking_refs) {
-            indicator.object_marking_refs = observable.object_marking_refs;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.labels) {
-            indicator.labels = observable.extensions[STIX_EXT_OCTI_SCO].labels;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.created_by_ref) {
-            indicator.created_by_ref = observable.extensions[STIX_EXT_OCTI_SCO].created_by_ref;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.external_references) {
-            indicator.external_references = observable.extensions[STIX_EXT_OCTI_SCO].external_references;
-          }
-          if (granted_refs) {
-            indicator.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(indicator);
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(indicator.id);
-          }
-          const relationBaseData = {
-            source_ref: indicator.id,
-            target_ref: observable.id,
-            relationship_type: RELATION_BASED_ON,
-          };
-          const relationStandardId = idGenFromData('relationship', relationBaseData);
-          const relationship = {
-            id: relationStandardId,
-            type: 'relationship',
-            ...relationBaseData,
-            object_marking_refs: observable.object_marking_refs ?? [],
-            created: now(),
-            modified: now(),
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: RELATION_BASED_ON,
-              },
-            },
-          } as StixRelation;
-          if (granted_refs) {
-            relationship.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(relationship);
-
-          // Resolve relationships in the bundle
-          const stixRelationshipsInBundle = bundle.objects.filter((r) => r.type === 'relationship') as StixRelation[];
-          const stixRelationships = stixRelationshipsInBundle.filter((r) => r.relationship_type === 'related-to'
-            && r.source_ref === baseData.id
-            && (
-              r.target_ref.startsWith('threat-actor')
-              || r.target_ref.startsWith('intrusion-set')
-              || r.target_ref.startsWith('campaign')
-              || r.target_ref.startsWith('malware')
-              || r.target_ref.startsWith('incident')
-              || r.target_ref.startsWith('tool')
-              || r.target_ref.startsWith('attack-pattern')
-            ));
-          for (let indexStixRelationships = 0; indexStixRelationships < stixRelationships.length; indexStixRelationships += 1) {
-            const stixRelationship = stixRelationships[indexStixRelationships] as StixRelation;
-            const relationIndicatesBaseData = {
-              source_ref: indicator.id,
-              target_ref: stixRelationship.target_ref,
-              relationship_type: RELATION_INDICATES,
-            };
-            const relationIndicatesStandardId = idGenFromData('relationship', relationIndicatesBaseData);
-            const relationshipIndicates = {
-              id: relationIndicatesStandardId,
-              type: 'relationship',
-              ...relationIndicatesBaseData,
-              object_marking_refs: observable.object_marking_refs ?? [],
-              created: now(),
-              modified: now(),
-              extensions: {
-                [STIX_EXT_OCTI]: {
-                  extension_type: 'property-extension',
-                  type: RELATION_INDICATES,
-                },
-              },
-            } as StixRelation;
-            if (granted_refs) {
-              relationshipIndicates.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-            }
-            objectsToPush.push(relationshipIndicates);
-          }
-          // Resolve relationships in database
-          if (isNotEmptyField(id)) {
-            const relationsOfObservables = await fullRelationsList(
-              context,
-              AUTOMATION_MANAGER_USER,
-              ABSTRACT_STIX_CORE_RELATIONSHIP,
-              {
-                fromOrToId: id,
-                toTypes: [
-                  ENTITY_TYPE_THREAT_ACTOR,
-                  ENTITY_TYPE_INTRUSION_SET,
-                  ENTITY_TYPE_CAMPAIGN,
-                  ENTITY_TYPE_MALWARE,
-                  ENTITY_TYPE_INCIDENT,
-                  ENTITY_TYPE_TOOL,
-                  ENTITY_TYPE_ATTACK_PATTERN,
-                ],
-                baseData: true,
-                indices: READ_RELATIONSHIPS_INDICES,
-              },
-            ) as StoreRelation[];
-            const idsToResolve = R.uniq(relationsOfObservables.map((r) => r.toId));
-            const elements = await stixLoadByIds(context, AUTOMATION_MANAGER_USER, idsToResolve);
-            for (let indexElements = 0; indexElements < elements.length; indexElements += 1) {
-              const element = elements[indexElements] as StixCoreObject;
-              const relationIndicatesBaseData = {
-                source_ref: indicator.id,
-                target_ref: element.id,
-                relationship_type: RELATION_INDICATES,
-              };
-              const relationIndicatesStandardId = idGenFromData('relationship', relationIndicatesBaseData);
-              const relationshipIndicates = {
-                id: relationIndicatesStandardId,
-                type: 'relationship',
-                ...relationIndicatesBaseData,
-                object_marking_refs: observable.object_marking_refs ?? [],
-                created: now(),
-                modified: now(),
-                extensions: {
-                  [STIX_EXT_OCTI]: {
-                    extension_type: 'property-extension',
-                    type: RELATION_INDICATES,
-                  },
-                },
-              } as StixRelation;
-              if (granted_refs) {
-                relationshipIndicates.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-              }
-              objectsToPush.push(relationshipIndicates);
-            }
-          }
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(relationship.id);
-          }
-        }
-      }
-    }
-    if (objectsToPush.length > 0) {
-      pushAll(bundle.objects, objectsToPush);
-      return { output_port: 'out', bundle: { ...bundle, objects: bundle.objects.map((n) => (n.id === baseData.id ? baseData : n)) } };
-    }
-    return { output_port: 'unmodified', bundle };
-  },
-};
-interface CreateObservableConfiguration {
-  all: boolean;
-  wrap_in_container: boolean;
-}
-const PLAYBOOK_CREATE_OBSERVABLE_COMPONENT_SCHEMA: JSONSchemaType<CreateObservableConfiguration> = {
-  type: 'object',
-  properties: {
-    all: { type: 'boolean', $ref: 'Create observables from all indicators in the bundle', default: false },
-    wrap_in_container: { type: 'boolean', $ref: 'If main entity is a container, wrap observables in container', default: false },
-  },
-  required: [],
-};
-const PLAYBOOK_CREATE_OBSERVABLE_COMPONENT: PlaybookComponent<CreateObservableConfiguration> = {
-  id: 'PLAYBOOK_CREATE_OBSERVABLE_COMPONENT',
-  name: 'Extract observables from indicator',
-  description: 'Create observables based on an indicator',
-  icon: 'observable',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }, { id: 'unmodified', type: 'out' }],
-  configuration_schema: PLAYBOOK_CREATE_OBSERVABLE_COMPONENT_SCHEMA,
-  schema: async () => PLAYBOOK_CREATE_OBSERVABLE_COMPONENT_SCHEMA,
-  executor: async ({ playbookNode, dataInstanceId, bundle }) => {
-    const { all, wrap_in_container } = playbookNode.configuration;
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle);
-    const indicators = [baseData];
-    if (all) {
-      pushAll(indicators, bundle.objects);
-    }
-    const { type: baseDataType } = baseData.extensions[STIX_EXT_OCTI];
-    const isBaseDataAContainer = isStixDomainObjectContainer(baseDataType);
-    const objectsToPush: StixObject[] = [];
-    for (let indexIndicator = 0; indexIndicator < indicators.length; indexIndicator += 1) {
-      const indicator = indicators[indexIndicator] as StixIndicator;
-      if (indicator.type === 'indicator') {
-        const observables = extractValidObservablesFromIndicatorPattern(indicator.pattern);
-        for (let indexObservable = 0; indexObservable < observables.length; indexObservable += 1) {
-          const observable = observables[indexObservable];
-          const description = indicator.description ?? `Simple observable of indicator {${indicator.name || indicator.pattern}}`;
-          const { score, granted_refs } = indicator.extensions[STIX_EXT_OCTI];
-          const observableData = {
-            ...R.dissoc('type', observable),
-            x_opencti_score: score,
-            x_opencti_description: description,
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: observable.type,
-              },
-              [STIX_EXT_OCTI_SCO]: {
-                extension_type: 'property-extension',
-                score,
-                description,
-              },
-            },
-          };
-          const observableStandardId = generateStandardId(observable.type, observableData);
-          const storeObservable = {
-            internal_id: generateInternalId(),
-            standard_id: observableStandardId,
-            entity_type: observable.type,
-            parent_types: getParentTypes(observable.type),
-            ...observableData,
-          } as StoreCommon;
-          const stixObservable = convertStoreToStix_2_1(storeObservable) as StixCyberObject;
-          if (indicator.object_marking_refs) {
-            stixObservable.object_marking_refs = indicator.object_marking_refs;
-          }
-          if (indicator.created_by_ref && stixObservable.extensions[STIX_EXT_OCTI_SCO]) {
-            stixObservable.extensions[STIX_EXT_OCTI_SCO].created_by_ref = indicator.created_by_ref;
-          }
-          if (indicator.labels && stixObservable.extensions[STIX_EXT_OCTI_SCO]) {
-            stixObservable.extensions[STIX_EXT_OCTI_SCO].labels = indicator.labels;
-          }
-          if (indicator.external_references && stixObservable.extensions[STIX_EXT_OCTI_SCO]) {
-            stixObservable.extensions[STIX_EXT_OCTI_SCO].external_references = indicator.external_references;
-          }
-          if (granted_refs) {
-            stixObservable.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(stixObservable);
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(stixObservable.id);
-          }
-          const relationBaseData = {
-            source_ref: indicator.id,
-            target_ref: stixObservable.id,
-            relationship_type: RELATION_BASED_ON,
-          };
-          const relationStandardId = idGenFromData('relationship', relationBaseData);
-          const relationship = {
-            id: relationStandardId,
-            type: 'relationship',
-            ...relationBaseData,
-            object_marking_refs: indicator.object_marking_refs ?? [],
-            created: now(),
-            modified: now(),
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: RELATION_BASED_ON,
-              },
-            },
-          } as StixRelation;
-          if (granted_refs) {
-            relationship.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(relationship);
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(relationship.id);
-          }
-        }
-      }
-    }
-    if (objectsToPush.length > 0) {
-      pushAll(bundle.objects, objectsToPush);
-      return { output_port: 'out', bundle: { ...bundle, objects: bundle.objects.map((n) => (n.id === baseData.id ? baseData : n)) } };
-    }
-    return { output_port: 'unmodified', bundle };
-  },
-};
-// endregion
 
 // @ts-expect-error TODO improve playbook types to avoid this
 export const PLAYBOOK_COMPONENTS: { [k: string]: PlaybookComponent<object> } = {
@@ -1870,7 +698,7 @@ export const PLAYBOOK_COMPONENTS: { [k: string]: PlaybookComponent<object> } = {
   [PLAYBOOK_INGESTION_COMPONENT.id]: PLAYBOOK_INGESTION_COMPONENT,
   [PLAYBOOK_MATCHING_COMPONENT.id]: PLAYBOOK_MATCHING_COMPONENT,
   [PLAYBOOK_CONNECTOR_COMPONENT.id]: PLAYBOOK_CONNECTOR_COMPONENT,
-  [PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT.id]: PLAYBOOK_UPDATE_KNOWLEDGE_COMPONENT,
+  [PLAYBOOK_MANIPULATE_KNOWLEDGE_COMPONENT.id]: PLAYBOOK_MANIPULATE_KNOWLEDGE_COMPONENT,
   [PLAYBOOK_CONTAINER_WRAPPER_COMPONENT.id]: PLAYBOOK_CONTAINER_WRAPPER_COMPONENT,
   [PLAYBOOK_SECURITY_COVERAGE_COMPONENT.id]: PLAYBOOK_SECURITY_COVERAGE_COMPONENT,
   [PLAYBOOK_SHARING_COMPONENT.id]: PLAYBOOK_SHARING_COMPONENT,
