+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbook-components-test.ts
@@ -1,14 +1,15 @@
 import { afterAll, beforeAll, describe, expect, it, vi } from 'vitest';
-import { PLAYBOOK_CONTAINER_WRAPPER_COMPONENT, PLAYBOOK_SHARING_COMPONENT } from '../../../src/modules/playbook/playbook-components';
 import type { StixBundle } from '../../../src/types/stix-2-1-common';
 import type { BasicStoreEntityOrganization } from '../../../src/modules/organization/organization-types';
 import { unSetOrganization, setOrganization } from '../../utils/testQueryHelper';
 import { PLATFORM_ORGANIZATION, TEST_ORGANIZATION } from '../../utils/testQuery';
 import { getOrganizationEntity } from '../../utils/domainQueryHelper';
-import { sharing_component_bundle } from './playbookComponents/playbook-sharing-component-test';
-import { container_wrapper_component_bundle } from './playbookComponents/playbook-container-wrapper-component-test';
+import { sharing_component_bundle } from './playbookComponents/sharing-component-test';
+import { container_wrapper_component_bundle } from './playbookComponents/container-wrapper-component-test';
 import { STIX_EXT_OCTI } from '../../../src/types/stix-2-1-extensions';
 import * as entrepriseEdition from '../../../src/enterprise-edition/ee';
+import { PLAYBOOK_CONTAINER_WRAPPER_COMPONENT } from '../../../src/modules/playbook/components/container-wrapper-component';
+import { PLAYBOOK_SHARING_COMPONENT } from '../../../src/modules/playbook/components/sharing-component';
 
 describe('playbook sharing component', () => {
   let externalOrganizationEntity: BasicStoreEntityOrganization;
