+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/access-restrictions-component-test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it, vi, beforeEach, afterEach, type MockInstance } from 'vitest';
 import type { StixBundle, StixObject, StixOpenctiExtension } from '../../../../src/types/stix-2-1-common';
 import { STIX_EXT_OCTI } from '../../../../src/types/stix-2-1-extensions';
-import { PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT } from '../../../../src/modules/playbook/playbook-components';
 import { ENTITY_TYPE_CONTAINER_REPORT } from '../../../../src/schema/stixDomainObject';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../../../../src/modules/organization/organization-types';
 import * as access from '../../../../src/utils/access';
 import * as authorizedMembers from '../../../../src/utils/authorizedMembers';
+import { PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT } from '../../../../src/modules/playbook/components/access-restrictions-component';
 
 describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
   const reportId = 'report--5f78a68b-2c4d-5e6f-beaa-7b987b0e7165';
