+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/connector-component-test.ts
