+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/sharing-component-test.ts
@@ -3,11 +3,11 @@ import type { StixBundle, StixObject, StixOpenctiExtension } from '../../../../s
 import type { BasicStoreObject } from '../../../../src/types/store';
 import type { StixId } from '../../../../src/types/stix-2-0-common';
 import { STIX_EXT_OCTI } from '../../../../src/types/stix-2-1-extensions';
-import { PLAYBOOK_SHARING_COMPONENT } from '../../../../src/modules/playbook/playbook-components';
 import { generateStandardId } from '../../../../src/schema/identifier';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../../../../src/modules/organization/organization-types';
 import * as middlewareLoader from '../../../../src/database/middleware-loader';
 import * as access from '../../../../src/utils/access';
+import { PLAYBOOK_SHARING_COMPONENT } from '../../../../src/modules/playbook/components/sharing-component';
 
 export const sharing_component_bundle = {
   id: '411628bf-745b-43f6-8194-cbe441edecfd',
