+++ opencti-platform/opencti-graphql/tests/03-integration/04-manager/playbooks/playbooks-test.ts
@@ -11,7 +11,7 @@ describe('buildStixTaskFromTaskTemplate tests', () => {
       base_type: 'ENTITY',
       confidence: 100,
       creator_id: [
-        '88ec0c6a-13ce-5e39-b486-354fe4a7084f'
+        '88ec0c6a-13ce-5e39-b486-354fe4a7084f',
       ],
       description: '',
       entity_type: 'Task-Template',
@@ -20,16 +20,16 @@ describe('buildStixTaskFromTaskTemplate tests', () => {
       name: 'read logs',
       parent_types: [
         'Basic-Object',
-        'Internal-Object'
+        'Internal-Object',
       ],
       'rel_template-task.internal_id': [
-        'c4f7df8d-c6a8-418e-a761-536587ec50c1'
+        'c4f7df8d-c6a8-418e-a761-536587ec50c1',
       ],
       sort: [
-        'task-template--c06283e4-4bc8-59ee-9788-86508394a63e'
+        'task-template--c06283e4-4bc8-59ee-9788-86508394a63e',
       ],
       standard_id: 'task-template--c06283e4-4bc8-59ee-9788-86508394a63e',
-      'template-task': 'c4f7df8d-c6a8-418e-a761-536587ec50c1'
+      'template-task': 'c4f7df8d-c6a8-418e-a761-536587ec50c1',
     } as unknown as BasicStoreEntityTaskTemplate;
 
     const container = {
@@ -38,22 +38,22 @@ describe('buildStixTaskFromTaskTemplate tests', () => {
         'extension-definition--ea279b3e-5c71-4632-ac08-831c66a786ba': {
           extension_type: 'new-sdo',
           id: '0b298142-29e8-4958-8b55-b07af2cb8870',
-          type: 'Case-Incident'
-        }
+          type: 'Case-Incident',
+        },
       },
       id: 'case-incident--e2a5b146-81f8-5d0d-8b58-6cfcd282c167',
       labels: [
-        'akira'
+        'akira',
       ],
       name: 'incident playbook',
       object_marking_refs: [
-        'marking-definition--89484dde-e3d2-547f-a6c6-d14824429eb1'
+        'marking-definition--89484dde-e3d2-547f-a6c6-d14824429eb1',
       ],
       object_refs: [
-        'incident--e97b1203-fa52-5803-8115-e4144a468189'
+        'incident--e97b1203-fa52-5803-8115-e4144a468189',
       ],
       spec_version: '2.1',
-      type: 'case-incident'
+      type: 'case-incident',
     } as unknown as StixContainer;
 
     const expectedTask = {
@@ -61,19 +61,19 @@ describe('buildStixTaskFromTaskTemplate tests', () => {
         'extension-definition--ea279b3e-5c71-4632-ac08-831c66a786ba': {
           extension_type: 'new-sdo',
           id: 'd50ba267-790b-46f6-b5ab-c6ed78389127',
-          type: 'Task'
-        }
+          type: 'Task',
+        },
       },
       id: 'task--af7b713a-a10e-5f5e-8a57-7341ab87a2c6',
       name: 'read logs',
       object_marking_refs: [
-        'marking-definition--89484dde-e3d2-547f-a6c6-d14824429eb1'
+        'marking-definition--89484dde-e3d2-547f-a6c6-d14824429eb1',
       ],
       object_refs: [
-        'case-incident--e2a5b146-81f8-5d0d-8b58-6cfcd282c167'
+        'case-incident--e2a5b146-81f8-5d0d-8b58-6cfcd282c167',
       ],
       spec_version: '2.1',
-      type: 'task'
+      type: 'task',
     };
     const task = buildStixTaskFromTaskTemplate(taskTemplate, container);
     expect(task.id).toEqual(expectedTask.id);
